requirements.txt
uvicorn
fastapi
black
flake8
pre-commit
python-multipart
numpy
numpy-stl
sympy